Conflicts:
	sport/src/main/java/org/cd/sport/action/SubjectSbOperAction.java
	sport/src/main/java/org/cd/sport/service/SubjectServiceImpl.java

+++ sport/src/main/java/org/cd/sport/service/UserServiceImpl.java
@@ -15,15 +15,16 @@ import org.cd.sport.exception.EntityNotFoundExcetion;
 import org.cd.sport.exception.NameIsExistException;
 import org.cd.sport.exception.ParameterIsWrongException;
 import org.cd.sport.support.UserSupport;
+import org.cd.sport.utils.AuthenticationUtils;
 import org.cd.sport.utils.Md5Util;
 import org.cd.sport.view.UserView;
+import org.cd.sport.vo.UserAuth;
 import org.cd.sport.vo.UserVo;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataAccessException;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.User;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
@@ -87,9 +88,15 @@ public class UserServiceImpl extends UserSupport implements UserService {
 		if (user == null) {
 			return null;
 		}
+		//TODO 优化
+		UserVo userVo = this.getVoById(user.getUserId());
 		// 用户权限集合
 		Collection<GrantedAuthority> grantedAuthority = this.getGrantedAuthority(user);
-		User userdetail = new User(user.getLoginName(), user.getPassword(), true, true, true, true, grantedAuthority);
+		UserAuth userdetail = new UserAuth(user.getLoginName(), user.getPassword(), true, true, true, true,
+				grantedAuthority);
+		userdetail.setUserId(user.getUserId());
+		userdetail.setUserName(user.getUserName());
+		userdetail.setOrgName(userVo.getOrgName());
 		return userdetail;
 	}
 
@@ -112,13 +119,16 @@ public class UserServiceImpl extends UserSupport implements UserService {
 	@Override
 	@Transactional
 	public boolean update(UserView user) {
-		this.validate(user);
+		this.validateUpdate(user);
 		UserDomain userDomain = this.userDao.findById(user.getUserId());
 		if (userDomain == null) {
 			throw new EntityNotFoundExcetion("数据不存在");
 		}
 		this.validLoginName(user.getLoginName(), userDomain);
+		String password = userDomain.getPassword();
 		BeanUtils.copyProperties(user, userDomain);
+		userDomain.setGender(Constants.User.parseGender(user.getGender()));
+		userDomain.setPassword(password);
 		this.userDao.update(userDomain);
 		return true;
 	}
@@ -136,26 +146,45 @@ public class UserServiceImpl extends UserSupport implements UserService {
 	@Override
 	@Transactional
 	public boolean resetPassword(String loginName) {
-		// 角色验证 TODO
+		UserVo userDomain = AuthenticationUtils.getUser();
 		UserDomain user = this.getByLoginName(loginName);
 		if (user == null) {
 			throw new EntityNotFoundExcetion("数据不存在");
 		}
-		user.setPassword(Md5Util.digestMD5(Constants.User.DEFAULT_PASSWORD));
-		this.userDao.update(user);
-		return true;
+		if (Constants.Role.hasOper(userDomain.getRole(), user.getRole()) && !Constants.Role.isAdmin(user.getRole())) {
+			user.setPassword(Md5Util.digestMD5(Constants.User.DEFAULT_PASSWORD));
+			this.userDao.update(user);
+			return true;
+		}
+		return false;
 	}
 
 	@Override
+	@Transactional
 	public boolean delete(String userId) {
-		// 角色验证 TODO
-		return this.userDao.deleteById(userId);
+		UserDomain user = this.getById(userId);
+		if (user == null) {
+			return false;
+		}
+		UserVo userDomain = AuthenticationUtils.getUser();
+		if (Constants.Role.hasOper(userDomain.getRole(), user.getRole()) && !Constants.Role.isAdmin(user.getRole())) {
+			return this.userDao.deleteById(userId);
+		}
+		return false;
 	}
 
 	@Override
+	@Transactional
 	public boolean delete(String[] userId) {
-		// 角色验证 TODO
-		return this.userDao.deleteById(userId);
+		if (userId == null) {
+			return false;
+		}
+		boolean result = true;
+		for (int i = 0; i < userId.length; i++) {
+			String uI = userId[i];
+			result &= this.delete(uI);
+		}
+		return result;
 	}
 
 	@Override
@@ -205,5 +234,21 @@ public class UserServiceImpl extends UserSupport implements UserService {
 			return this.getTotal(role);
 		}
 		return this.userDao.count(role, name);
+	}
+
+	@Override
+	public UserVo getVoById(String userId) {
+		if (StringUtils.isBlank(userId)) {
+			return null;
+		}
+		return this.userDao.findVoById(userId);
+	}
+
+	@Override
+	public UserVo getVoByLoginName(String loginName) {
+		if (StringUtils.isBlank(loginName)) {
+			return null;
+		}
+		return this.userDao.findVoByLoginName(loginName);
 	}
 }
